Renaming test_helper.rb to blockly_interpreter_test_helper.rb so that it doesn't conflict with any test_helper files in the apps using this gem
