+++ test/test_helper.rb
@@ -1,6 +1,6 @@
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
 require 'blockly_interpreter'
-require 'blockly_interpreter/test_helper'
+require 'blockly_interpreter/blockly_interpreter_test_helper'
 
 require 'minitest/spec'
 require 'minitest/reporters'
